+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Delete comment at cursor position with `<leader>rd` keymap and `:CodeReviewDeleteComment` command
+- Auto-copy formatted comment to clipboard when adding new comments with `comment.auto_copy_on_add` option
 
 ## [0.2.0] - 2025-07-01